Scrap the RGB gradient concept.
This commit does away with the RGB color gradient idea where the newer,
non NULL characters in a CharStream would have a darker color than the
older ones. In place of color, we are indicating whether a non NULL Char
is first in a stream of Chars. This allows the user of ScreenBuffer to
take special action with the first character that appears in stream.